Add support for specifying console with runtest
Some targets (s390) configure a console which does not exist
as device in /dev, but is listed in /proc/consoles. The console
can be specified on the command line and will work. However, it
can not be used to run a getty on it.

Add support for specifying such consoles in the configuration
file. Also, add two fallbacks:
- if /proc/consoles does not point to a valid device,
  try to read the device from /sys/class/tty/console/active
- if that does not yield a console, try using /dev/console directly

Only bail out and execute run.sh directly if all attempts to find
a working console fail.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>